build: fix the build of `unified-bridge` on zkvm (#7)
Pruned a bunch of unused dependencies from the `unified-bridge` crate.

The problematic one is `sp1-sdk`. It cannot be built for zkvm since it
pulls in IO-related functionality through `tokio`. That is not supported
inside zkvm.

Also used this as an opportunity to prune a bunch of other unused
dependencies.