refactor: wrap `alloy_primitives::Address` and give data types specific `serde` names (#40)
## Motivation

In order to avoid further storage compatibility regressions,
`serde-reflection` can be employed to capture a snapshot of the encoding
format as understood by serde. The package, however, has a number of
limitations around handling generics and distinct types with identical
names defined in different modules (a common occurrence with `Error`).
This

## Additions and Changes

* The `alloy_primitives::Address` is wrapped and exposed as
`agglayer_primitives::Address` with a specific serde name.
* Some other types have been given specific serde names to avoid name
clashes.

## Breaking changes

Hopefully none, serde names at type level do not factor into the
serialization format of `json`, `toml`, or `bincode`.

## PR Checklist:

- [ ] I have performed a self-review of my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added or updated tests that comprehensively prove my change
is effective or that my feature works