build: introduce Agglayer ELF builder (#47)
# Description

This introduces a zkvm ELF builder based on `build.rs` script, fully
integrated with the standard Rust build system.

It's controlled by the `AGGLAYER_BUILD` environment variable which
offers multiple build modes, most notably:

* `AGGLAYER_BUILD=update` - Auto-update the zkvm ELF from source
(requires docker to be installed).
* `AGGLAYER_BUILD=cached` - Use cached committed zkvm ELF (default, no
special requirements).

In either case, it is not necessary to have succinct toolchain or cargo
prove installed on the build machine. These are provided by the docker
image when needed.

For more detail, have a look at the top-level module comment for
`agglayer_build::zkvm_build`.

## Breaking changes

Build scripts are not aware of artifacts written to the source tree.
Therefore, it is no longer legal to just
`include_bytes!("path/to/elf/in/repo")`. Instead, the crate that
performs the zkvm ELF build using `build.rs` should export the ELF
binary as a Rust `const`ant byte array. Should the ELF binary be
required in any other crate, it needs to import the elf crate as a
dependency. This avoids potential build race conditions.