feat: introduce an agglayer-errors crate (#106)
First part of https://github.com/agglayer/agglayer/issues/838 : with
this we'll have an easy-to-use framework to deal with eyre errors, as
well as the always-useful `InternalError` error marker.

Though, a question for the reviewer: Should we actually have an
`InternalError` marker, or should we just say that any error that has no
other known marker is an internal error? I started introducing it, but
re-reading my code before submitting the PR makes me think this might be
a step backwards, and instead we should maybe just remove that and say
"anything not otherwise-defined is an internal error".